# Conflicts:
#	src/main/java/de/bmoth/parser/ErrorListener.java
#	src/main/java/de/bmoth/parser/ast/SemanticAstCreator.java
#	src/main/java/de/bmoth/parser/ast/nodes/NumberNode.java

+++ src/main/java/de/bmoth/parser/ast/nodes/NumberNode.java
@@ -2,16 +2,18 @@ package de.bmoth.parser.ast.nodes;
 
 import org.antlr.v4.runtime.ParserRuleContext;
 
+import java.math.BigInteger;
+
 public class NumberNode extends ExprNode {
 
-    private final int value;
+    private final BigInteger value;
 
-    public NumberNode(ParserRuleContext ctx, int value) {
+    public NumberNode(ParserRuleContext ctx, BigInteger value) {
         super(ctx);
         this.value = value;
     }
 
-    public int getValue() {
+    public BigInteger getValue() {
         return value;
     }
